Monkey patches removes number parsing from yargs
None of the arguments we are going to pass will be numbers (I think?) so
remove that from yargs entirely. Had to format the code and try to find
the lines that was responsible for it, I think I did. `S()` checks if
the string matches a number regex, and if so, passes it to Number()
constructor. Well I just removed the Number constructor altogether.

Ultimately I'd like use the config `parse-positional-numbers` but that
isn't available in whatever version this is.

This is needed because if we have a commit like 1e99999, that'll be
treated as exponential / scientific notation, which ends up getting
rounded up to Infinity.